+++ assembler/src/debruijn/path_extend/split_graph_pair_info.hpp
@@ -241,9 +241,9 @@ private:
                 size_t kmer_distance = read_distance
                         + mapping_edge_2.second.initial_range.end_pos
                         - mapping_edge_1.second.initial_range.start_pos;
-                int edge_distance = kmer_distance
-                        + mapping_edge_1.second.mapped_range.start_pos
-                        - mapping_edge_2.second.mapped_range.end_pos;
+                int edge_distance = (int) kmer_distance
+                        + (int) mapping_edge_1.second.mapped_range.start_pos
+                        - (int) mapping_edge_2.second.mapped_range.end_pos;
 
                 basket_index.AddPairInfo(mapping_edge_1.first,
                         mapping_edge_1.second.mapped_range.start_pos,
@@ -323,7 +323,7 @@ public:
                         DEBUG("Thread number " << omp_get_thread_num() << " is going to increase its limit by " << coeff << " times, current limit is " << limit);
                     }
                     buffer_pi[i]->Clear();
-                    limit = coeff * limit;
+                    limit = (size_t) (coeff * (double) limit);
                 }
             }
             DEBUG("Thread number " << omp_get_thread_num() << " finished");
